+++ src/utils/createKey.ts
@@ -1,4 +1,4 @@
 /** @ignore */
-export function createKey(keyLength: number): string {
+export function createKey(keyLength?: number): string {
 	return Math.random().toString(36).substr(2, keyLength);
 }